Delivery of the first version of the Equipment Inventory system.
The organization of pages was improved, positioning login as the initial page at the moment of starting the system.
Improved pop-up messages.
TO DO: Save images to generate PDF delivery of equipment.
In reports the CPU section add the filtering of mouse, keyboard, monitor.
In reports improve the filtering of fields